add setuptools config, move examples
