+++ .gitignore
@@ -1,4 +1,5 @@
 __pycache__
+*.egg-info
 .vscode/*
 
 # emacs stuff
