--- requirements.txt
@@ -1,3 +1,0 @@
-PyYAML==6.0
-drawsvg==2.1.1
-pyyaml-include==1.3
