+++ setup.py
@@ -1,0 +1,5 @@
+from setuptools import setup
+
+# a dummy setup.py is needed for --editable instals
+
+setup()
