+++ src/examples/dxcore/extract_variants.py
