+++ src/examples/dxcore/family/dxcore_page.json
