+++ src/examples/dxcore/family/junk/dxcore.py
