+++ src/examples/dxcore/template/dxcore_label.py
