+++ src/examples/dxcore/template/dxcore_page.py
