+++ src/examples/dxcore/template/function_label.py
