+++ src/examples/dxcore/template/junk/dxcore_colors.json
