+++ src/examples/dxcore/template/junk/dxcore_colors.py
