+++ src/examples/dxcore/template/packages.py
