+++ src/examples/dxcore/template/page.py
