+++ src/examples/dxcore/template/tengo_functions.json
