+++ src/examples/dxcore/variant/DD20.json
