+++ src/examples/dxcore/variant/DD20_qfp.json
