+++ src/examples/dxcore/variant/DD20_soic.json
