+++ src/examples/dxcore/variant/DD20_sop.json
