+++ src/examples/dxcore/variant/junk/DD20.py
