+++ src/examples/dxcore/variant/junk/functions.json
