+++ src/examples/dxcore/variant/junk/junk.json
