+++ src/examples/tengo10/pinouts/dxcore/dd14_20/qfp20h.yaml
