+++ src/examples/tengo10/pinouts/dxcore/template/package_qfp.py
